--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,11 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
-	"fmt"
+	"bufio"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go"
 )
 
 type User struct {
@@ -19,48 +18,36 @@ type User struct {
 	Address  string
 }
 
+var json = jsoniter.ConfigCompatibleWithStandardLibrary
+
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
+	return newCountDomains(r, domain)
 }
 
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
+type Emails struct {
+	Email string
 }
 
-func countDomains(u users, domain string) (DomainStat, error) {
+func newCountDomains(r io.Reader, domain string) (DomainStat, error) {
 	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
+	b := strings.Builder{}
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		var user Emails
+		line := scanner.Bytes()
+		if err := json.Unmarshal(line, &user); err != nil {
 			return nil, err
 		}
+		b.WriteRune('.')
+		b.WriteString(domain)
+		afterAt := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+		matched := strings.HasSuffix(afterAt, b.String())
+		b.Reset()
 
 		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+			result[afterAt]++
 		}
 	}
 	return result, nil