@xkdgo 2021-12-08
before
goos: linux
goarch: amd64
pkg: github.com/xkdgo/otus_golang_homeworks/hw10_program_optimization
cpu: Intel(R) Pentium(R) CPU G4560 @ 3.50GHz
BenchmarkGetDomainStatFile-4   	      15	 761517893 ns/op	306750565 B/op	 3145431 allocs/op
BenchmarkGetDomainStatFile-4   	      15	 763549649 ns/op	306772889 B/op	 3145434 allocs/op
BenchmarkGetDomainStatFile-4   	      14	 764653779 ns/op	306784915 B/op	 3145432 allocs/op
BenchmarkGetDomainStatFile-4   	      15	 759947346 ns/op	306812050 B/op	 3145435 allocs/op
BenchmarkGetDomainStatFile-4   	      15	 760232715 ns/op	306819917 B/op	 3145438 allocs/op
PASS
ok  	github.com/xkdgo/otus_golang_homeworks/hw10_program_optimization	64.301s

after
goos: linux
goarch: amd64
pkg: github.com/xkdgo/otus_golang_homeworks/hw10_program_optimization
cpu: Intel(R) Pentium(R) CPU G4560 @ 3.50GHz
BenchmarkGetDomainStatFile-4   	      51	 231900724 ns/op	18645274 B/op	 1801034 allocs/op
BenchmarkGetDomainStatFile-4   	      50	 232022906 ns/op	18645302 B/op	 1801034 allocs/op
BenchmarkGetDomainStatFile-4   	      50	 232173686 ns/op	18645356 B/op	 1801034 allocs/op
BenchmarkGetDomainStatFile-4   	      50	 232748492 ns/op	18645359 B/op	 1801034 allocs/op
BenchmarkGetDomainStatFile-4   	      50	 232180965 ns/op	18645338 B/op	 1801034 allocs/op
PASS
ok  	github.com/xkdgo/otus_golang_homeworks/hw10_program_optimization	59.475s

delta
name                 old time/op    new time/op    delta
GetDomainStatFile-4     762ms ± 0%     232ms ± 0%  -69.53%

name                 old alloc/op   new alloc/op   delta
GetDomainStatFile-4     307MB ± 0%      19MB ± 0%  -93.92%

name                 old allocs/op  new allocs/op  delta
GetDomainStatFile-4     3.15M ± 0%     1.80M ± 0%  -42.74%